requirements.txt
flask
flask-smorest
flask-migrate
flask-sqlalchemy
sqlalchemy
psycopg2-binary
python-dotenv